Separate RandomChain fuzzing utils into package code so I can read coverage info for it
